+++ R/findInFiles.R
@@ -23,7 +23,7 @@
 #' @return A tibble if \code{output="tibble"}, otherwise a
 #'   \code{htmlwidget} object.
 #'
-#' @import htmlwidgets
+#' @importFrom htmlwidgets createWidget
 #' @importFrom stringr str_split_fixed
 #' @importFrom crayon strip_style
 #' @importFrom tibble tibble
@@ -133,7 +133,7 @@ findInFiles <- function(
   }
 
   # create widget
-  htmlwidgets::createWidget(
+  createWidget(
     name = "findInFiles",
     x = x,
     width = NULL,
@@ -210,111 +210,3 @@ FIF2dataframe <- function(fif){
   as.data.frame(tbl)
 }
 
-#' @title Shiny bindings for \code{findInFiles}
-#'
-#' @description Output and render functions for using \code{findInFiles} within
-#'   Shiny applications and interactive Rmd documents.
-#'
-#' @param outputId output variable to read from
-#' @param width,height a valid CSS unit (like \code{"100\%"},
-#'   \code{"400px"}, \code{"auto"}) or a number, which will be coerced to a
-#'   string and have \code{"px"} appended
-#' @param expr an expression that generates a '\code{\link{findInFiles}}' widget
-#' @param env the environment in which to evaluate \code{expr}
-#' @param quoted logical, whether \code{expr} is a quoted expression (with
-#'   \code{quote()})
-#'
-#' @return \code{FIFOutput} returns an output element that can be included in a
-#'   Shiny UI definition, and \code{renderFIF} returns a
-#'   \code{shiny.render.function} object that can be included in a Shiny server
-#'   definition.
-#'
-#' @name findInFiles-shiny
-#'
-#' @import htmlwidgets
-#' @export
-#'
-#' @examples library(findInFiles)
-#' library(shiny)
-#'
-#' onKeyDown <- HTML(
-#'   'function onKeyDown(event) {',
-#'   '  var key = event.which || event.keyCode;',
-#'   '  if(key === 13) {',
-#'   '    Shiny.setInputValue(',
-#'   '      "pattern", event.target.value, {priority: "event"}',
-#'   '    );',
-#'   '  }',
-#'   '}'
-#' )
-#'
-#' ui <- fluidPage(
-#'   tags$head(tags$script(onKeyDown)),
-#'   br(),
-#'   sidebarLayout(
-#'     sidebarPanel(
-#'       selectInput(
-#'         "ext", "Extension",
-#'         choices = c("R", "js", "css")
-#'       ),
-#'       tags$div(
-#'         class = "form-group shiny-input-container",
-#'         tags$label(
-#'           class = "control-label",
-#'           "Pattern"
-#'         ),
-#'         tags$input(
-#'           type = "text",
-#'           class = "form-control",
-#'           onkeydown = "onKeyDown(event);",
-#'           placeholder = "Press Enter when ready"
-#'         )
-#'       ),
-#'       numericInput(
-#'         "depth", "Depth (set -1 for unlimited depth)",
-#'         value = 0, min = -1, step = 1
-#'       ),
-#'       checkboxInput(
-#'         "wholeWord", "Whole word"
-#'       ),
-#'       checkboxInput(
-#'         "ignoreCase", "Ignore case"
-#'       )
-#'     ),
-#'     mainPanel(
-#'       FIFOutput("results")
-#'     )
-#'   )
-#' )
-#'
-#'
-#' server <- function(input, output){
-#'
-#'   output[["results"]] <- renderFIF({
-#'     req(input[["pattern"]])
-#'     findInFiles(
-#'       ext = isolate(input[["ext"]]),
-#'       pattern = input[["pattern"]],
-#'       depth = isolate(input[["depth"]]),
-#'       wholeWord = isolate(input[["wholeWord"]]),
-#'       ignoreCase = isolate(input[["ignoreCase"]])
-#'     )
-#'   })
-#'
-#' }
-#'
-#' if(interactive()){
-#'   shinyApp(ui, server)
-#' }
-FIFOutput <- function(outputId, width = "100%", height = "400px"){
-  htmlwidgets::shinyWidgetOutput(
-    outputId, "findInFiles", width, height, package = "findInFiles"
-  )
-}
-
-#' @rdname findInFiles-shiny
-#' @export
-renderFIF <- function(expr, env = parent.frame(), quoted = FALSE){
-  if(!quoted){ expr <- substitute(expr) } # force quoted
-  htmlwidgets::shinyRenderWidget(expr, FIFOutput, env, quoted = TRUE)
-}