+++ inst/shinyApp/global.R
@@ -1,0 +1,11 @@
+library(findInFiles)
+library(shiny)
+library(shinyAce)
+library(shinyWidgets)
+library(shinyjqui)
+library(shinyFiles)
+library(shinyvalidate)
+library(fs)
+
+roots <- c(wd = getOption("FIFWD"), getVolumes()())
+